Throwing in a quick change to show wrong way handiling
I know you're not supposed to use this forbidden statement, it's just a quick thing to show using turnAround and doing the method again handles going the wrong way and I have to get going. Follow Gradient works a lot better now, although I'm not sure how to handle if it goes straight onto the circle? If it follows the line, how can it know if it ran into a triangle or a circle?