Follow Gradient and Feed Wander behaviors finished
Could use some improvements here and there, but it seems to be good. Moved them to tasks. The feed task has some trouble staying on the patch and I just can't figure out a way to improve it, could use some help with that. But hopefully we can start integrating this into main soon.