Finilizing Gradient Follow Behavior
Added in another check for leftLumence spike that was causing problems. Looking for the gradient is now handled by it's own task that runs in the background as well. With Integration into the main code being the next step, I'm not sure how the task program flow will work, but we'll figure it out.